+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## [0.0.6](../../releases/tag/v0.0.6) - Unreleased
+## [0.0.7](../../releases/tag/v0.0.7) - Unreleased
+
+### Fixes
+
+- selector handling for `RETRY_CSS_SELECTORS` in `_handle_blocked_request` in `BeautifulSoupCrawler`
+- selector handling in `enqueue_links` in `BeautifulSoupCrawler`
+
+## [0.0.6](../../releases/tag/v0.0.6) - 2024-06-25
 
 ### Adds